requirements.txt
pyyaml
redis
schedule
ffmpeg-python